Feat: 랜덤 사각형 생성 객체 구현 - Plane은 생성된 사각형들을 가지고 관리하는 기능만으로 한정 - 랜덤한 사각형을 생성하고 전달하는 기능은 RandomRectangleMaker 객체에게 부여
